refactored createAction
