--- src/createAction.js
@@ -1,49 +1,0 @@
-var _ = require('./utils'),
-    Reflux = require('../src'),
-    Keep = require('./Keep'),
-    allowed = {preEmit:1,shouldEmit:1};
-
-/**
- * Creates an action functor object. It is mixed in with functions
- * from the `PublisherMethods` mixin. `preEmit` and `shouldEmit` may
- * be overridden in the definition object.
- *
- * @param {Object} definition The action object definition
- */
-module.exports = function(definition) {
-
-    definition = definition || {};
-
-    for(var a in Reflux.ActionMethods){
-        if (!allowed[a] && Reflux.PublisherMethods[a]) {
-            throw new Error("Cannot override API method " + a +
-                " in Reflux.ActionMethods. Use another method name or override it on Reflux.PublisherMethods instead."
-            );
-        }
-    }
-
-    for(var d in definition){
-        if (!allowed[d] && Reflux.PublisherMethods[d]) {
-            throw new Error("Cannot override API method " + d +
-                " in action creation. Use another method name or override it on Reflux.PublisherMethods instead."
-            );
-        }
-    }
-
-    var context = _.extend({
-        eventLabel: "action",
-        emitter: new _.EventEmitter(),
-        _isAction: true
-    }, Reflux.PublisherMethods, Reflux.ActionMethods, definition);
-
-    var functor = function() {
-        functor[functor.sync?"trigger":"triggerAsync"].apply(functor, arguments);
-    };
-
-    _.extend(functor,context);
-
-    Keep.createdActions.push(functor);
-
-    return functor;
-
-};
